Revised Email Tag implementation - RCML/Core.
