+++ restcomm/restcomm.interpreter/src/main/java/org/mobicents/servlet/restcomm/interpreter/ConfVoiceInterpreter.java
@@ -17,9 +17,10 @@ import org.apache.http.HttpStatus;
 import org.apache.http.NameValuePair;
 import org.apache.http.message.BasicNameValuePair;
 import org.joda.time.DateTime;
-import org.mobicents.servlet.restcomm.email.CreateEmailService;
-import org.mobicents.servlet.restcomm.email.api.EmailRequest;
-import org.mobicents.servlet.restcomm.email.api.Mail;
+import org.mobicents.servlet.restcomm.email.api.CreateEmailService;
+import org.mobicents.servlet.restcomm.email.api.EmailService;
+import org.mobicents.servlet.restcomm.email.EmailRequest;
+import org.mobicents.servlet.restcomm.email.Mail;
 import org.mobicents.servlet.restcomm.cache.DiskCache;
 import org.mobicents.servlet.restcomm.cache.DiskCacheRequest;
 import org.mobicents.servlet.restcomm.cache.DiskCacheResponse;
@@ -308,7 +309,7 @@ public class ConfVoiceInterpreter extends UntypedActor {
 
             @Override
             public Actor create() throws Exception {
-                final CreateEmailService.Builder builder = CreateEmailService.builder();
+                final CreateEmailService builder = new EmailService();
                 builder.CreateEmailSession(configuration);
                 EMAIL_SENDER=builder.getUser();
                 return builder.build();
