+++ restcomm/restcomm.ussd/src/main/java/org/mobicents/servlet/restcomm/ussd/interpreter/UssdInterpreter.java
@@ -50,9 +50,10 @@ import org.joda.time.DateTime;
 import org.mobicents.servlet.restcomm.dao.CallDetailRecordsDao;
 import org.mobicents.servlet.restcomm.dao.DaoManager;
 import org.mobicents.servlet.restcomm.dao.NotificationsDao;
-import org.mobicents.servlet.restcomm.email.CreateEmailService;
-import org.mobicents.servlet.restcomm.email.api.EmailRequest;
-import org.mobicents.servlet.restcomm.email.api.Mail;
+import org.mobicents.servlet.restcomm.email.api.CreateEmailService;
+import org.mobicents.servlet.restcomm.email.api.EmailService;
+import org.mobicents.servlet.restcomm.email.EmailRequest;
+import org.mobicents.servlet.restcomm.email.Mail;
 import org.mobicents.servlet.restcomm.entities.CallDetailRecord;
 import org.mobicents.servlet.restcomm.entities.Notification;
 import org.mobicents.servlet.restcomm.entities.Sid;
@@ -293,7 +294,7 @@ public class UssdInterpreter extends UntypedActor {
 
             @Override
             public Actor create() throws Exception {
-                final CreateEmailService.Builder builder = CreateEmailService.builder();
+                final CreateEmailService  builder = new EmailService();
                 builder.CreateEmailSession(configuration);
                 EMAIL_SENDER=builder.getUser();
                 return builder.build();
